feat(location): implement location queries and enhance API endpoint
- Added new location query functions for finding locations by name and slug, and generating unique slugs.
- Refactored the API endpoint for creating locations to utilize the new query functions, improving code organization and readability.
- Updated error handling in the dashboard add page to provide more specific error messages.
- Enhanced the location schema to include type definitions for better type safety.
